Setup anchor pairs in markFeatureWriter
